+++ plotly/CHANGELOG.md
@@ -9,10 +9,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 ### Changed
+- Plotly moved on top of VisionForge (in the VisionForge repository)
+- Geo module merged into core
 
 ### Deprecated
 
 ### Removed
+- Own HtmlFragment replaced by VisionForge Html handling
 
 ### Fixed